refactor(assets): thin __init__.py, move logic to manager.py (C3, C4, C8, C10)
- assets/__init__.py reduced from 282 to 42 lines (thin exports only)
- All implementation moved to assets/manager.py
- Fixed docstring: clarified assets are NOT bundled, meshes are
  downloaded on-demand from robot_descriptions/GitHub (C4)
- Added comment explaining lazy import of download_assets (C8)
- download.py now re-exports from tools/download_assets.py (C10)

This addresses the core structural concern about __init__.py being
too heavy (AGENTS.md Convention #3).
